Changed the end plate to corners
Trying to make the design easier to build, smaller, but still rigid.
Now the support plate is solid, but only the corners are reinforced.

Todo: move the Z motors to the top... to match the other motors, and to
give better support to the single motor side.

Todo: Add mounting holes for the various electronics bits.