Added an idler
I neglected to add the second idler for the belt crossover... whoops.
Luckily I didn't cut before noticing.

A todo is to move the crossover inside the endcap, but for the first
build I'd rather have it out where I can keep an eye on it... and figure
out how to actually build it.