Added tabs to the end plates, and center mount for the support plate
The center of the wall plates was a bit wobbly for my taste.  Of course, that was with only the corners attached - but still.

Moved the Z stage inwards a couple mm - I really need my m8 flanged bearings to come in so I can mount the Z screws and see how that's going to work.
My guess is I'll add a couple small pincher screws to clamp the flanged bearings in.

Paul